Fixes the type of model `l2_block` field `block_id` from i64 to i32.

+++ src/db/models/rollup_state/l2_block.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 
 #[derive(sqlx::FromRow, Serialize, Debug, Clone)]
 pub struct L2Block {
-    pub block_id: i64,
+    pub block_id: i32,
     pub new_root: String,
     pub witness: serde_json::Value,
     pub created_time: TimestampDbType,
